Added a new TravelInfoSectionItem, and some other small tweaks from CR
