+++ src/stacks-hierarchy/Root_TabNavigatorStack/TabNav_DashboardStack/Dashboard_RootScreen/components/CompactFareContracts.tsx
@@ -1,7 +1,7 @@
 import {Button} from '@atb/components/button';
 import {useFirestoreConfigurationContext} from '@atb/configuration/FirestoreConfigurationContext';
 import {CompactFareContractInfo} from '@atb/fare-contracts/CompactFareContractInfo';
-import {getFareContractInfoDetails} from '@atb/fare-contracts/FareContractInfo';
+import {getFareContractInfoDetails} from '@atb/fare-contracts/FareContractInfoDetails';
 import {StyleSheet, useThemeContext} from '@atb/theme';
 import {
   DashboardTexts,
